make features work for logged in users

+++ Chitra/checkout/views.py
@@ -8,33 +8,36 @@ from django.http import HttpResponseRedirect
 from django.contrib import messages
 
 def order(request):
-    if request.method == 'POST':
-        chars = "[]"
-        posts = request.POST['posts']
-        if (posts):
-            phone_number = request.POST['phone_number']
-            zone = request.POST['zone']
-            city = request.POST['city']
-            state = request.POST['state']
-            zip_code = request.POST['zip_code']
-            for c in chars:
-                posts = posts.replace(c, '')
-                products = posts.split(',')
-            if request.user.is_authenticated:
-                user = request.user
-                for i in products:
-                    product_details = all_posts.objects.get(id=int(i))
-                    order = checkout(product=product_details, user=user, phone_number=phone_number, zone=zone, state=state, city=city, zip_code=zip_code )
-                    order.save()
-            response = HttpResponseRedirect('/')
-            response.delete_cookie('items')
-            messages.success(request, "Your order has been placed!!")
-            return response
+    if (request.user.is_authenticated):
+        if request.method == 'POST':
+            chars = "[]"
+            posts = request.POST['posts']
+            if (posts):
+                phone_number = request.POST['phone_number']
+                zone = request.POST['zone']
+                city = request.POST['city']
+                state = request.POST['state']
+                zip_code = request.POST['zip_code']
+                for c in chars:
+                    posts = posts.replace(c, '')
+                    products = posts.split(',')
+                if request.user.is_authenticated:
+                    user = request.user
+                    for i in products:
+                        product_details = all_posts.objects.get(id=int(i))
+                        order = checkout(product=product_details, user=user, phone_number=phone_number, zone=zone, state=state, city=city, zip_code=zip_code )
+                        order.save()
+                response = HttpResponseRedirect('/')
+                response.delete_cookie('items')
+                messages.success(request, "Your order has been placed!!")
+                return response
+            else:
+                messages.error(request, "Oops!! It seems there isn't any product to checkout.")
+                return render(request, 'checkout.html')
         else:
-            messages.error(request, "Oops!! It seems there isn't any product to checkout.")
             return render(request, 'checkout.html')
     else:
-        return render(request, 'checkout.html')
+        return redirect('/')
 
 @csrf_exempt
 def items_display(request):
